move GenerativeBayes to naive_bayes module
